internal/database: Refactor ImportCount
Make ImportCount return the number of distinct importing packages in the
importers table. This fixes an issue where multiple versions of an
importing package can inflate the import count, and also removes the
need to maintain a separate import_counts table.

The database needs to be migrated with the following SQL script:

BEGIN;

DROP TABLE import_counts;

CREATE FUNCTION import_count(import_path text)
RETURNS bigint
LANGUAGE SQL
AS $$
SELECT COUNT(DISTINCT importer_path) FROM importers WHERE import_path = $1;
$$;

ALTER TABLE importers
ADD FOREIGN KEY (importer_path, importer_version)
REFERENCES packages (import_path, version) ON DELETE CASCADE;

COMMIT;

+++ gddo-server/template.go
@@ -87,7 +87,7 @@ type Package struct {
 	Updated     time.Time
 	SubPackages []database.Package
 	Meta        *source.Meta
-	ImportCount int
+	ImportCount int64
 	allExamples []*texample
 }
